+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # Version 3.5.10
 
 * Added Manual Changelog
+* Changed Latex Processing Unit to make crafted dry rubber instead of tiny ones with increased fluid latex req